+++ src/components/BulkEditList/BulkEditListSidebar/BulkEditListSidebar.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { useHistory, useLocation } from 'react-router-dom';
 
 import { ButtonGroup } from '@folio/stripes/components';
@@ -12,7 +13,7 @@ import { QueryTab } from './QueryTab/QueryTab';
 import { LogsTab } from './LogsTab/LogsTab';
 import { useSearchParams } from '../../../hooks/useSearchParams';
 
-export const BulkEditListSidebar = () => {
+export const BulkEditListSidebar = ({ isBuildQueryButtonDisabled }) => {
   const history = useHistory();
   const location = useLocation();
   const { hasLogViewPerms, hasQueryPerms } = useBulkPermissions();
@@ -45,10 +46,14 @@ export const BulkEditListSidebar = () => {
       )}
 
       {/* QUERY TAB */}
-      {isQuery && hasQueryPerms && <QueryTab />}
+      {isQuery && hasQueryPerms && <QueryTab isBuildQueryButtonDisabled={isBuildQueryButtonDisabled} />}
 
       {/* LOGS TAB */}
       {isLogs && <LogsTab />}
     </>
   );
+};
+
+BulkEditListSidebar.propTypes = {
+  isBuildQueryButtonDisabled: PropTypes.bool,
 };